+++ src/Model/Uninstall/Uninstall.php
@@ -38,9 +38,13 @@ class Uninstall extends Model {
         delete_option($slug . '-version');
         delete_option($slug . '-check-updates');
         delete_option($slug . '-modules-states');
-        // For site options in Multisite
+
+        delete_option('wp_register_files');
+
         delete_site_option($slug . '-version');
         delete_site_option($slug . '-check-updates');  
-        delete_site_option($slug . '-modules-states');  
+        delete_site_option($slug . '-modules-states');
+
+        delete_site_option('wp_register_files');
     }
 }